[feat] UserMainScreen 및 데모 리포지토리 구현

+++ core/model/src/main/java/com/ssavice/model/service/SearchResultItem.kt
@@ -11,7 +11,7 @@ data class SearchResultItem(
     val minimumMember: Int,
     val currentMember: Int,
     val basePrice: Long,
-    val discountRatio: Double,
+    val discountRatio: Int,
     val discountedPrice: Long,
     val deadLine: Date,
     val latitude: Double,
